created index page on OOP
